requirements.txt
agno==1.0.6
altair==5.5.0
anaconda-anon-usage @ file:///C:/b/abs_e8r_zga7xy/croot/anaconda-anon-usage_1732732454901/work
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
boltons @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/boltons_1699480450092/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.5.1
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi @ file:///C:/b/abs_90yq4lmu83/croot/cffi_1726856448345/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
conda @ file:///C:/b/abs_3bs4tei4e9/croot/conda_1733841834375/work
conda-anaconda-telemetry @ file:///C:/b/abs_b5wbfhyt0i/croot/conda-anaconda-telemetry_1733260551861/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_bdz9vbvbh2/croot/conda-package-streaming_1731366449946/work
contourpy==1.3.1
cryptography @ file:///C:/b/abs_e2lzchf4i6/croot/cryptography_1732130411942/work
cycler==0.12.1
defusedxml==0.7.1
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
docstring_parser==0.16
duckduckgo_search==7.3.0
feedparser==6.0.11
filelock==3.17.0
fonttools==4.55.8
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
libmambapy @ file:///C:/b/abs_7fjjcok9fe/croot/mamba-split_1732896315069/work/libmambapy
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
menuinst @ file:///C:/b/abs_cabvzz4p7a/croot/menuinst_1731364936347/work
multitasking==0.0.11
narwhals==1.25.1
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.2
oauthlib==3.2.2
openai==1.61.1
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
primp==0.11.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycosat @ file:///C:/b/abs_5csdern___/croot/pycosat_1714513102923/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml @ file:///C:/b/abs_0cunwx_ww6/croot/ruamel.yaml_1727980181547/work
ruamel.yaml.clib @ file:///C:/b/abs_5fk8zi6n09/croot/ruamel.yaml.clib_1727769837359/work
seaborn==0.13.2
setuptools==75.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
streamlit==1.42.0
tabulate==0.9.0
tenacity==9.0.0
tldextract==5.1.3
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm @ file:///C:/b/abs_77wju137gk/croot/tqdm_1724853945787/work
truststore @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/truststore_1701881385424/work
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
watchdog==6.0.0
webencodings==0.5.1
wheel==0.44.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
yfinance==0.2.54
youtube-transcript-api==0.6.3
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
